+++ src/main/java/com/supermartijn642/benched/seat/SeatEntity.java
@@ -5,6 +5,7 @@ import com.supermartijn642.benched.blocks.BenchBlock;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.protocol.Packet;
+import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.network.NetworkHooks;
@@ -55,7 +56,7 @@ public class SeatEntity extends Entity {
     }
 
     @Override
-    public Packet<?> getAddEntityPacket(){
+    public Packet<ClientGamePacketListener> getAddEntityPacket(){
         return NetworkHooks.getEntitySpawningPacket(this);
     }
 }