+++ changelog.md
@@ -1,2 +1,2 @@
 ### Benched 1.2.2
-- Initial release for Minecraft 1.19.3
+- Initial release for Minecraft 1.20