+++ changelog.md
@@ -1,3 +1,6 @@
+### Benched 1.2.2a
+- Added Mangrove bench
+
 ### Benched 1.2.2
 - Updated to core library 1.1
 - Added all the benches to the 'Decorations' creative tab